Waypoints
I have made the waypoints work with a player unit, the system will use the units pos as start for the search then get a random end one, it then moves the player unit to the end pos after finding the path moving along the path. There is somethings I want to change and make better but the system is at a working point. also added a Game Manager script in this to keep some global variables that I will need all classes to have access to.