Turn system and added pro text.
I have made a turn system where it will check the initiative of the units puts them in the right order then runs the game. Player and AI will move the unit that is the current turn and the press end turn button, if they have not move the unit it will display text telling them that they have not moved and will not end the unit's turn. also added a text pack to the project for better text.