Unit addition and changes
Add the Ai versions of the player units, made a setup function for each for easier adding more units later and less repetition. Also added a new attack damage variable and new health and max health variable. Also change the damage function to now check if unit is dead and to remove from order list and set unit as not active.