Added tile effects
Made it so the sand tile reduces the range of ranged units, makes tank and warrior take 25% more damage and rogues have a 25% change to be hidden and doge damage. Mountain makes ranged units have more range and tanks and warriors have a 25% to block damage or rogues have a 25% to doge damage and water makes it so the unit can't attack.